Util split out Compile script
