+++ ui/guest/src/classes/ContentTypeHelper.ts
@@ -14,8 +14,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { createLookupTable, retrieveProperty } from '../util';
 import { ContentType, ContentTypeField } from '../models/ContentType';
+import { createLookupTable, retrieveProperty } from '../utils/object';
 
 export class ContentTypeHelper {
