+++ ui/guest/src/components/DropMarker.tsx
@@ -16,20 +16,17 @@
 
 import React, { useEffect, useRef, useState } from 'react';
 import {
-  findClosestRect,
   foo,
-  forEach,
-  getDropMarkerPosition,
-  getInRectStats,
   HORIZONTAL,
-  notNullOrUndefined,
-  splitRect,
   VERTICAL,
   X_AXIS,
   Y_AXIS
-} from '../util';
+} from '../utils/util';
 import { Coordinates } from '../models/Positioning';
 import { DropZone, Record } from '../models/InContextEditing';
+import { notNullOrUndefined } from '../utils/object';
+import { forEach } from '../utils/array';
+import { findClosestRect, getDropMarkerPosition, getInRectStats, splitRect } from '../utils/dom';
 
 interface DropMarkerProps {
   over: Record;
