+++ ui/guest/src/hooks.ts
@@ -14,12 +14,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { foo, isNullOrUndefined, notNullOrUndefined } from './util';
+import { foo} from './utils/util';
 import { useContext, useEffect, useRef, useState } from 'react';
 import contentController from './classes/ContentController';
 import { GuestContext } from './components/GuestContext';
 import { skip } from 'rxjs/operators';
 import { ContentInstance } from './models/ContentInstance';
+import { isNullOrUndefined, notNullOrUndefined } from './utils/object';
 
 const fooEvents = {
   ref: foo,
